+++ src/test/java/ladder/domain/ladder/MoveDirectionTest.java
@@ -1,5 +1,6 @@
 package ladder.domain.ladder;
 
+import ladder.util.MoveDirection;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
